@@ -20,35 +20,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.1.1",
-    "@angular/cdk": "^13.1.2",
-    "@angular/common": "^13.1.1",
-    "@angular/compiler": "^13.1.1",
-    "@angular/core": "^13.1.1",
-    "@angular/forms": "^13.1.1",
-    "@angular/material": "^13.1.2",
-    "@angular/platform-browser": "^13.1.1",
-    "@angular/platform-browser-dynamic": "^13.1.1",
-    "@angular/router": "^13.1.1",
-    "@enigmatry/angular-building-blocks": "^1.0.110",
-    "@enigmatry/formly-extensions": "^1.0.96",
-    "@ngx-formly/core": "^5.10.22",
-    "@ngx-formly/material": "^5.10.22",
-    "nswag": "^13.12.1",
+    "@angular/animations": "^14.1.3",
+    "@angular/cdk": "^14.1.3",
+    "@angular/common": "^14.1.3",
+    "@angular/compiler": "^14.1.3",
+    "@angular/core": "^14.1.3",
+    "@angular/forms": "^14.1.3",
+    "@angular/material": "^14.1.3",
+    "@angular/platform-browser": "^14.1.3",
+    "@angular/platform-browser-dynamic": "^14.1.3",
+    "@angular/router": "^14.1.3",
+    "@enigmatry/angular-building-blocks": "^1.0.126",
+    "@enigmatry/formly-extensions": "^1.0.126",
+    "@ngx-formly/core": "^5.12.6",
+    "@ngx-formly/material": "^5.12.6",
+    "nswag": "^13.16.1",
     "rxjs": "^6.6.7",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.1.3",
-    "@angular/cli": "^13.1.3",
-    "@angular/compiler-cli": "^13.1.2",
-    "@angular/localize": "^13.1.1",
-    "@enigmatry/eslint-config": "^1.0.110",
-    "@enigmatry/stylelint-config": "^1.0.110",
+    "@angular-devkit/build-angular": "^14.1.3",
+    "@angular/cli": "^14.1.3",
+    "@angular/compiler-cli": "^14.1.3",
+    "@angular/localize": "^14.1.3",
+    "@enigmatry/eslint-config": "^1.0.126",
+    "@enigmatry/stylelint-config": "^1.0.126",
     "@types/jasmine": "^3.8.1",
     "@types/node": "^16.4.3",
-    "eslint": "^8.11.0",
+    "eslint": "^8.22.0",
     "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.3.4",
@@ -60,6 +60,6 @@
     "protractor": "~7.0.0",
     "stylelint": "^14.6.0",
     "ts-node": "~10.4.0",
-    "typescript": "4.5.4"
+    "typescript": "4.7.4"
   }
 }
