@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { FieldType } from '@ngx-formly/material/form-field';
 
 @Component({
@@ -7,5 +7,5 @@ import { FieldType } from '@ngx-formly/material/form-field';
   styleUrls: ['./formly-readonly-boolean.component.scss']
 })
 export class ReadonlyBooleanComponent extends FieldType {
-  formControl: FormControl;
+  formControl: UntypedFormControl;
 }
