@@ -1,6 +1,6 @@
 //----------------------
 // <auto-generated>
-//     Generated using the NSwag toolchain v13.15.10.0 (NJsonSchema v10.6.10.0 (Newtonsoft.Json v13.0.0.0)) (http://NSwag.org)
+//     Generated using the NSwag toolchain v13.16.1.0 (NJsonSchema v10.7.2.0 (Newtonsoft.Json v13.0.0.0)) (http://NSwag.org)
 // </auto-generated>
 //----------------------
 
@@ -30,7 +30,7 @@ export interface IUsersClient {
      * Creates or updates
      * @param command User data
      */
-    post(command: UserCreateOrUpdateCommand): Observable<GetUserDetailsResponse>;
+    post(command: CreateOrUpdateUserCommand): Observable<GetUserDetailsResponse>;
     /**
      * Get user for given id
      * @param id Id
@@ -130,7 +130,7 @@ export class UsersClient implements IUsersClient {
      * Creates or updates
      * @param command User data
      */
-    post(command: UserCreateOrUpdateCommand): Observable<GetUserDetailsResponse> {
+    post(command: CreateOrUpdateUserCommand): Observable<GetUserDetailsResponse> {
         let url_ = this.baseUrl + "/Users";
         url_ = url_.replace(/[?&]$/, "");
 
@@ -971,12 +971,12 @@ export interface IProblemDetails {
     instance?: string | undefined;
 }
 
-export class UserCreateOrUpdateCommand implements IUserCreateOrUpdateCommand {
+export class CreateOrUpdateUserCommand implements ICreateOrUpdateUserCommand {
     id?: string | undefined;
     userName?: string;
     name?: string;
 
-    constructor(data?: IUserCreateOrUpdateCommand) {
+    constructor(data?: ICreateOrUpdateUserCommand) {
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
@@ -993,9 +993,9 @@ export class UserCreateOrUpdateCommand implements IUserCreateOrUpdateCommand {
         }
     }
 
-    static fromJS(data: any): UserCreateOrUpdateCommand {
+    static fromJS(data: any): CreateOrUpdateUserCommand {
         data = typeof data === 'object' ? data : {};
-        let result = new UserCreateOrUpdateCommand();
+        let result = new CreateOrUpdateUserCommand();
         result.init(data);
         return result;
     }
@@ -1009,7 +1009,7 @@ export class UserCreateOrUpdateCommand implements IUserCreateOrUpdateCommand {
     }
 }
 
-export interface IUserCreateOrUpdateCommand {
+export interface ICreateOrUpdateUserCommand {
     id?: string | undefined;
     userName?: string;
     name?: string;
